requirements.txt
djangorestframework
drf-spectacular
requests
python-dotenv
django_filter
pillow